requirements.txt
torch
torchvision
ipykernel
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
optuna
jupyter
fastapi[standard]
uvicorn[standard]
pydantic
torchaudio
obspy
seisbench
h5py
tqdm
wandb
einops
accelerate
ema-pytorch
alembic
appdirs
asttokens
brotli
charset-normalizer
click
colorama
colorlog
comm
pillow
pip
platformdirs
python-dotenv
gunicorn
einops
tensorflow
tensorflow-cpu
tensorflow-gpu